# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a battleships game server.",
        "title": "Swagger Example API",
        "contact": {
            "name": "Shamil Garatuev",
            "email": "garatuev@gmail.com"
        },
        "license": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/clear": {
            "post": {
                "description": "clear the battlefield",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "BattleField"
                ],
                "summary": "clear the battlefield",
                "responses": {
                    "200": {},
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/create-matrix": {
            "post": {
                "description": "create new battlefield with provided size",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "BattleField"
                ],
                "summary": "create new battlefield",
                "parameters": [
                    {
                        "description": "createParams",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/battlefield.CreateFieldRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/battlefield.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/battlefield.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ship": {
            "post": {
                "description": "add ships to battlefield\ninput params should be like this:\n\"A1 B2,C4 C6,E7 F8\" where first coordinate is one corner of ship, second - other.\nships can be square or rectangular\nships can't be placed on top of each other and near each other.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Ships"
                ],
                "summary": "add ships to battlefield",
                "parameters": [
                    {
                        "description": "coordinates",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/battlefield.AddShipsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/battlefield.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/battlefield.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shot": {
            "post": {
                "description": "make a shot to provided coordinate\nexample: \"A1\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "BattleField"
                ],
                "summary": "make a shot to provided coordinate",
                "parameters": [
                    {
                        "description": "shot coordinates",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/battlefield.ShotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/battlefield.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "battlefield.AddShipsRequest": {
            "type": "object",
            "properties": {
                "coords": {
                    "type": "string"
                }
            }
        },
        "battlefield.CreateFieldRequest": {
            "type": "object",
            "properties": {
                "range": {
                    "type": "integer"
                }
            }
        },
        "battlefield.HTTPError": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                }
            }
        },
        "battlefield.ShotRequest": {
            "type": "object",
            "properties": {
                "coord": {
                    "type": "string"
                }
            }
        }
    }
}